# Conflicts:
#	app/src/main/java/com/habit/tracker/data/source/remote/interceptor/RetryInterceptor.kt
#	app/src/main/java/com/habit/tracker/di/ApplicationComponent.kt
#	app/src/main/java/com/habit/tracker/di/ViewModelModule.kt
#	app/src/main/java/com/habit/tracker/presentation/view/RegistrationFragment.kt

+++ app/src/main/java/com/habit/tracker/presentation/view/MapFragment.kt
@@ -157,7 +157,7 @@ class MapFragment : Fragment() {
                     })
                     val pointAnnotationOptions: PointAnnotationOptions =
                         PointAnnotationOptions()
-                            .withPoint(Point.fromLngLat(geo.long, geo.lat))
+                            .withPoint(Point.fromLngLat(geo.long.toDouble(), geo.lat.toDouble()))
                             .withIconImage(b)
                     pointAnnotationManager.create(pointAnnotationOptions)
                 }
